enhancement(select component): added relationship handler
