+++ src/features/users/CreateUserForm.jsx
@@ -48,14 +48,13 @@ function CreateUserForm({ onCloseModal }) {
         <Select
           name="role"
           onChange={roleField.onChange}
-          isMulti={false}
           isSearchable={true}
           defaultValue={[roleField.value]}
-          required={true}
-          disabled={isCreating}
-          options={Object.entries(USER_ROLES).map(([key, o]) => {
+          defaultOptions={Object.entries(USER_ROLES).map(([key, o]) => {
             return { value: key, label: o.label };
           })}
+          required={true}
+          disabled={isCreating}
         />
       </FormRow>
